# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Sistem Penjualan iPhone",
        "description": "Dokumentasi API untuk pengelolaan data iPhone (CRUD)",
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Registrasi User",
                "operationId": "97c1032e86d3fa8d070b25f7844f8e92",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registrasi berhasil"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login User",
                "operationId": "0117816bfd9c8c9aedb040ba3fb75439",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login berhasil"
                    }
                }
            }
        },
        "/api/iphones": {
            "get": {
                "tags": [
                    "Iphones"
                ],
                "summary": "Ambil semua data iPhone",
                "operationId": "5e42e6e97e53059809a1bed6c0dadfc4",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data"
                    }
                }
            },
            "post": {
                "tags": [
                    "Iphones"
                ],
                "summary": "Tambah data iPhone",
                "operationId": "e3a137ebd19a249963a1a44e28c2db85",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama",
                                    "storage",
                                    "harga",
                                    "stok"
                                ],
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "example": "iPhone 13"
                                    },
                                    "storage": {
                                        "type": "string",
                                        "example": "128GB"
                                    },
                                    "harga": {
                                        "type": "integer",
                                        "example": 12000000
                                    },
                                    "stok": {
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil ditambahkan"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/iphones/{id}": {
            "get": {
                "tags": [
                    "Iphones"
                ],
                "summary": "Detail iPhone",
                "operationId": "35178f07d652e30d7f807146a2e1a35e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil"
                    }
                }
            },
            "put": {
                "tags": [
                    "Iphones"
                ],
                "summary": "Update data iPhone",
                "operationId": "c6bb5ed6c898a434cf5e8b45b032936e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama": {
                                        "type": "string"
                                    },
                                    "storage": {
                                        "type": "string"
                                    },
                                    "harga": {
                                        "type": "integer"
                                    },
                                    "stok": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil diupdate"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Iphones"
                ],
                "summary": "Hapus data iPhone",
                "operationId": "4495070d2545fcdb5430fcb415bce523",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Berhasil dihapus"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Authentication"
        },
        {
            "name": "Iphones",
            "description": "Iphones"
        }
    ]
}